requirements.txt
discord.py>=2.3.2
google-generativeai>=0.4.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.28
psycopg2-binary>=2.9.9
